+++ tests/test_freeze_dynamic.py
@@ -19,14 +19,11 @@ import fixtures
 
 import capa.main
 import capa.rules
-import capa.helpers
 import capa.features.file
 import capa.features.insn
 import capa.features.common
 import capa.features.freeze
-import capa.features.basicblock
 import capa.features.extractors.null
-import capa.features.extractors.base_extractor
 from capa.features.address import Address, AbsoluteVirtualAddress
 from capa.features.extractors.base_extractor import (
     SampleHashes,
@@ -47,14 +44,19 @@ EXTRACTOR = capa.features.extractors.null.NullDynamicFeatureExtractor(
     ),
     global_features=[],
     file_features=[
-        (AbsoluteVirtualAddress(0x402345), capa.features.common.Characteristic("embedded pe")),
+        (
+            AbsoluteVirtualAddress(0x402345),
+            capa.features.common.Characteristic("embedded pe"),
+        ),
     ],
     processes={
         ProcessAddress(pid=1): capa.features.extractors.null.ProcessFeatures(
             name="explorer.exe",
             features=[],
             threads={
-                ThreadAddress(ProcessAddress(pid=1), tid=1): capa.features.extractors.null.ThreadFeatures(
+                ThreadAddress(
+                    ProcessAddress(pid=1), tid=1
+                ): capa.features.extractors.null.ThreadFeatures(
                     features=[],
                     calls={
                         DynamicCallAddress(
@@ -63,11 +65,21 @@ EXTRACTOR = capa.features.extractors.null.NullDynamicFeatureExtractor(
                             name="CreateFile(12)",
                             features=[
                                 (
-                                    DynamicCallAddress(thread=ThreadAddress(ProcessAddress(pid=1), tid=1), id=1),
+                                    DynamicCallAddress(
+                                        thread=ThreadAddress(
+                                            ProcessAddress(pid=1), tid=1
+                                        ),
+                                        id=1,
+                                    ),
                                     capa.features.insn.API("CreateFile"),
                                 ),
                                 (
-                                    DynamicCallAddress(thread=ThreadAddress(ProcessAddress(pid=1), tid=1), id=1),
+                                    DynamicCallAddress(
+                                        thread=ThreadAddress(
+                                            ProcessAddress(pid=1), tid=1
+                                        ),
+                                        id=1,
+                                    ),
                                     capa.features.insn.Number(12),
                                 ),
                             ],
@@ -78,7 +90,12 @@ EXTRACTOR = capa.features.extractors.null.NullDynamicFeatureExtractor(
                             name="WriteFile()",
                             features=[
                                 (
-                                    DynamicCallAddress(thread=ThreadAddress(ProcessAddress(pid=1), tid=1), id=2),
+                                    DynamicCallAddress(
+                                        thread=ThreadAddress(
+                                            ProcessAddress(pid=1), tid=1
+                                        ),
+                                        id=2,
+                                    ),
                                     capa.features.insn.API("WriteFile"),
                                 ),
                             ],
@@ -100,15 +117,18 @@ def test_null_feature_extractor():
     th = ThreadHandle(ThreadAddress(ProcessAddress(pid=1), tid=1), None)
 
     assert addresses(EXTRACTOR.get_processes()) == [ProcessAddress(pid=1)]
-    assert addresses(EXTRACTOR.get_threads(ph)) == [ThreadAddress(ProcessAddress(pid=1), tid=1)]
+    assert addresses(EXTRACTOR.get_threads(ph)) == [
+        ThreadAddress(ProcessAddress(pid=1), tid=1)
+    ]
     assert addresses(EXTRACTOR.get_calls(ph, th)) == [
         DynamicCallAddress(thread=ThreadAddress(ProcessAddress(pid=1), tid=1), id=1),
         DynamicCallAddress(thread=ThreadAddress(ProcessAddress(pid=1), tid=1), id=2),
     ]
 
-    rules = capa.rules.RuleSet([
-        capa.rules.Rule.from_yaml(
-            textwrap.dedent("""
+    rules = capa.rules.RuleSet(
+        [
+            capa.rules.Rule.from_yaml(
+                textwrap.dedent("""
                     rule:
                         meta:
                             name: create file
@@ -119,8 +139,9 @@ def test_null_feature_extractor():
                             - and:
                                 - api: CreateFile
                     """)
-        ),
-    ])
+            ),
+        ]
+    )
     capabilities = capa.main.find_capabilities(rules, EXTRACTOR)
     assert "create file" in capabilities.matches
 
@@ -131,14 +152,20 @@ def compare_extractors(a: DynamicFeatureExtractor, b: DynamicFeatureExtractor):
     assert addresses(a.get_processes()) == addresses(b.get_processes())
     for p in a.get_processes():
         assert addresses(a.get_threads(p)) == addresses(b.get_threads(p))
-        assert sorted(set(a.extract_process_features(p))) == sorted(set(b.extract_process_features(p)))
+        assert sorted(set(a.extract_process_features(p))) == sorted(
+            set(b.extract_process_features(p))
+        )
 
         for t in a.get_threads(p):
             assert addresses(a.get_calls(p, t)) == addresses(b.get_calls(p, t))
-            assert sorted(set(a.extract_thread_features(p, t))) == sorted(set(b.extract_thread_features(p, t)))
+            assert sorted(set(a.extract_thread_features(p, t))) == sorted(
+                set(b.extract_thread_features(p, t))
+            )
 
             for c in a.get_calls(p, t):
-                assert sorted(set(a.extract_call_features(p, t, c))) == sorted(set(b.extract_call_features(p, t, c)))
+                assert sorted(set(a.extract_call_features(p, t, c))) == sorted(
+                    set(b.extract_call_features(p, t, c))
+                )
 
 
 def test_freeze_str_roundtrip():